+++ mapie/classification.py
@@ -29,7 +29,6 @@ from .utils import (
     check_verbose,
     compute_quantiles,
     fit_estimator,
-    get_true_label_position
 )
 
 
@@ -710,6 +709,37 @@ class MapieClassifier(BaseEstimator, ClassifierMixin):
         )
         return conf_score
 
+    def _get_true_label_position(
+        self,
+        y_pred_proba: NDArray,
+        y: NDArray
+    ) -> NDArray:
+        """Return the sorted position of the true label in the
+        prediction
+
+        Parameters
+        ----------
+        y_pred_proba : NDArray of shape (n_samples, n_calsses)
+            Model prediction.
+        y : NDArray of shape (n_samples)
+            Labels.
+
+        Returns
+        -------
+        NDArray of shape (n_samples, 1)
+            Position of the true label in the prediction.
+        """
+        index = np.argsort(
+                np.fliplr(np.argsort(y_pred_proba, axis=1))
+            )
+        position = np.take_along_axis(
+            index,
+            y.reshape(-1, 1),
+            axis=1
+        )
+
+        return position
+
     def _get_last_included_proba(
         self,
         y_pred_proba: NDArray,
@@ -967,7 +997,7 @@ class MapieClassifier(BaseEstimator, ClassifierMixin):
             self.y_pred_proba_raps = estimator.predict_proba(
                 self.X_raps
             )
-            self.position_raps = get_true_label_position(
+            self.position_raps = self._get_true_label_position(
                 self.y_pred_proba_raps,
                 self.y_raps
             )
@@ -1039,7 +1069,7 @@ class MapieClassifier(BaseEstimator, ClassifierMixin):
             # Here we reorder the labels by decreasing probability
             # and get the position of each label from decreasing probability
 
-            self.conformity_scores_ = get_true_label_position(
+            self.conformity_scores_ = self._get_true_label_position(
                 y_pred_proba,
                 y
             )
